Dockerize app (#12)
## Why is this change necessary?
Need things to be containerized for deployment


 ## How does this change address the issue?
Creates a workflow to build images. Builds frontend and backend images
in CI and creates a downloadable artifact of them.
Creates Dockerfile for frontend and backend. Creates a batch script to
run on the windows machine to complete the deployment and bring up the
stack


 ## What side effects does this change have?
None


 ## How is this change tested?
in devcontainer and on my local laptop in two different downstream repos
(with and without backend)


 ## Other
removed some things from tsconfig.json that are already being managed by
the default nuxt config