Split E2E testing (#19)
## Why is this change necessary?
Lots of flakiness when trying to run dockerized and regular server E2E
tests as part of the same suite (conflicts with ports, failed builds...)


 ## How does this change address the issue?
Splits it to run as two separate suites. Same tests, just docker-compose
vs not. Removes the `$fetch` test of static HTML, which seems to provide
minimal value.


 ## What side effects does this change have?
Extra command to run tests


 ## How is this change tested?
In CI and locally in downstream repos